fix: move shared types to runtime
