+++ src/runtime/composables/index.ts
@@ -23,7 +23,7 @@ import { getLocale, getLocales, getComposer } from '../compatibility'
 import type { Ref } from 'vue'
 import type { Locale } from 'vue-i18n'
 import type { RouteLocation, RouteLocationNormalizedLoaded, RouteLocationRaw, Router } from 'vue-router'
-import type { I18nHeadMetaInfo, I18nHeadOptions, SeoAttributesOptions } from '../../types'
+import type { I18nHeadMetaInfo, I18nHeadOptions, SeoAttributesOptions } from '../shared-types'
 import type { HeadParam } from '../utils'
 
 export * from 'vue-i18n'
