+++ src/runtime/plugins/i18n.ts
@@ -29,8 +29,8 @@ import { createLogger } from 'virtual:nuxt-i18n-logger'
 import type { NuxtI18nPluginInjections } from '../injections'
 import type { Locale, I18nOptions } from 'vue-i18n'
 import type { NuxtApp } from '#app'
-import type { LocaleObject } from '../../types'
-import type { ModulePublicRuntimeConfig } from '../../module'
+import type { LocaleObject } from '../shared-types'
+import type { I18nPublicRuntimeConfig } from '../shared-types'
 
 // from https://github.com/nuxt/nuxt/blob/2466af53b0331cdb8b17c2c3b08675c5985deaf3/packages/nuxt/src/core/templates.ts#L152
 type Decorate<T extends Record<string, unknown>> = { [K in keyof T as K extends string ? `$${K}` : never]: T[K] }
@@ -59,7 +59,7 @@ export default defineNuxtPlugin<NuxtI18nPluginInjections>({
 
     // Fresh copy per request to prevent reusing mutated options
     const runtimeI18n = {
-      ...(nuxtContext.$config.public.i18n as ModulePublicRuntimeConfig['i18n']),
+      ...(nuxtContext.$config.public.i18n as I18nPublicRuntimeConfig),
       defaultLocale: defaultLocaleDomain
     }
     // @ts-expect-error type incompatible
